refactor(cache-schema): polish schema API and derive tests
