+++ crates/ssa-cache-schema/src/impls/mod.rs
@@ -4,13 +4,13 @@ mod tuples;
 
 use std::marker::PhantomData;
 
-use crate::{CacheSchema, ProductStyle, SchemaWriter};
+use crate::{CacheSchema, EmptyProductStyle, SchemaWriter};
 
 impl<T: CacheSchema> CacheSchema for Option<T> {
     fn write_schema(w: &mut SchemaWriter) {
         w.enum_begin("Option");
         w.variant_begin(0, "None");
-        w.product_style(ProductStyle::Unit);
+        w.empty_product_style(EmptyProductStyle::Unit);
         w.variant_end();
         w.variant_begin(1, "Some");
         w.field_begin(0, None);
@@ -67,7 +67,7 @@ where
 
 impl<T: ?Sized> CacheSchema for PhantomData<T> {
     fn write_schema(w: &mut SchemaWriter) {
-        w.primitive("PhantomData");
+        w.leaf("PhantomData");
     }
 }
